requirements.txt
tqdm>=4.19.5
pytest>=3.4.1
Pillow>=5.0.0
numpy>=1.14.1
